+++ src/main/java/de/maxhenkel/corpse/entities/CorpseRenderState.java
@@ -1,26 +1,14 @@
 package de.maxhenkel.corpse.entities;
 
-import net.minecraft.client.renderer.entity.SkeletonRenderer;
-import net.minecraft.client.renderer.entity.player.PlayerRenderer;
+import net.minecraft.client.renderer.entity.state.AvatarRenderState;
 import net.minecraft.client.renderer.entity.state.EntityRenderState;
-import net.minecraft.client.renderer.entity.state.PlayerRenderState;
 import net.minecraft.client.renderer.entity.state.SkeletonRenderState;
-
-import javax.annotation.Nullable;
 
 public class CorpseRenderState extends EntityRenderState {
 
     public float yRot;
-
-    @Nullable
-    public PlayerRenderer playerRenderer;
-    @Nullable
-    public PlayerRenderState playerRenderState;
-    @Nullable
-    public SkeletonRenderer skeletonRenderer;
-    @Nullable
-    public SkeletonRenderState skeletonRenderState;
-
+    public AvatarRenderState playerRenderState = new AvatarRenderState();
+    public SkeletonRenderState skeletonRenderState = new SkeletonRenderState();
     public boolean skeleton;
 
 }