+++ changelog.md
@@ -1,1 +1,1 @@
-- Improved corpse spawn position
+- Updated to 1.21.9